+++ spark/v3.1/spark/src/main/scala/com/netease/arctic/spark/sql/execution/ExtendedArcticStrategy.scala
@@ -23,6 +23,7 @@ import com.netease.arctic.spark.sql.catalyst.plans._
 import com.netease.arctic.spark.table.ArcticSparkTable
 import com.netease.arctic.spark.writer.WriteMode
 import org.apache.spark.sql.arctic.catalyst.ExpressionHelper
+import org.apache.spark.sql.arctic.execution.CreateArcticTableAsSelectExec
 import org.apache.spark.sql.catalyst.analysis.{NamedRelation, ResolvedTable}
 import org.apache.spark.sql.catalyst.expressions.PredicateHelper
 import org.apache.spark.sql.catalyst.plans.logical.{CreateTableAsSelect, DescribeRelation, LogicalPlan}
