+++ spark/v3.1/spark/src/main/scala/org/apache/spark/sql/arctic/execution/WritingSparkTask.scala
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
+ *  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,24 +16,26 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.execution.datasources.v2
+package org.apache.spark.sql.arctic.execution
 
 import com.netease.arctic.spark.writer.RowLevelWriter
 import org.apache.spark.executor.CommitDeniedException
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.connector.write.{DataWriter, DataWriterFactory}
+import org.apache.spark.sql.connector.write.{DataWriter, DataWriterFactory, WriterCommitMessage}
 import org.apache.spark.util.Utils
 import org.apache.spark.{SparkEnv, TaskContext}
 
 trait WritingSparkTask[W <: DataWriter[InternalRow]] extends Logging with Serializable {
 
   protected def writeFunc(writer: RowLevelWriter[InternalRow], row: InternalRow): Unit
+
   def run(
-           writerFactory: DataWriterFactory,
-           context: TaskContext,
-           iter: Iterator[InternalRow],
-           useCommitCoordinator: Boolean): DataWritingSparkTaskResult = {
+    writerFactory: DataWriterFactory,
+    context: TaskContext,
+    iter: Iterator[InternalRow],
+    useCommitCoordinator: Boolean
+  ): DataWritingSparkTaskResult = {
     val stageId = context.stageId()
     val stageAttempt = context.stageAttemptNumber()
     val partId = context.partitionId()
@@ -75,15 +77,21 @@ trait WritingSparkTask[W <: DataWriter[InternalRow]] extends Logging with Serial
 
       DataWritingSparkTaskResult(count, msg)
 
-    })(catchBlock = {
-      // If there is an error, abort this writer
-      logError(s"Aborting commit for partition $partId (task $taskId, attempt $attemptId, " +
-        s"stage $stageId.$stageAttempt)")
-      dataWriter.abort()
-      logError(s"Aborted commit for partition $partId (task $taskId, attempt $attemptId, " +
-        s"stage $stageId.$stageAttempt)")
-    }, finallyBlock = {
-      dataWriter.close()
-    })
+    })(
+      catchBlock = {
+        // If there is an error, abort this writer
+        logError(s"Aborting commit for partition $partId (task $taskId, attempt $attemptId, " +
+          s"stage $stageId.$stageAttempt)")
+        dataWriter.abort()
+        logError(s"Aborted commit for partition $partId (task $taskId, attempt $attemptId, " +
+          s"stage $stageId.$stageAttempt)")
+      }, finallyBlock = {
+        dataWriter.close()
+      })
   }
 }
+
+case class DataWritingSparkTaskResult(
+  numRows: Long,
+  writerCommitMessage: WriterCommitMessage
+)
